[API] Move transform_shapes from geometry to lmd_reader
