+++ src/dvpio/read/shapes/__init__.py
@@ -1,4 +1,3 @@
-from .geometry import transform_shapes
-from .lmd_reader import read_lmd
+from .lmd_reader import read_lmd, transform_shapes
 
 __all__ = ["read_lmd", "transform_shapes"]
