+++ src/dvpio/read/shapes/lmd_reader.py
@@ -4,7 +4,66 @@ import numpy as np
 import shapely
 from spatialdata.models import PointsModel, ShapesModel
 
-from .geometry import transform_shapes
+from .geometry import apply_affine_transformation, compute_affine_transformation
+
+
+def transform_shapes(
+    shapes: ShapesModel,
+    calibration_points_target: PointsModel,
+    calibration_points_source: PointsModel,
+    precision: int = 3,
+) -> ShapesModel:
+    """Apply coordinate transformation to shapes based on calibration points from a target and a source
+
+    Computes transformation between source and target coordinates.
+
+    Parameters
+    ----------
+    shapes
+        Shapes in source coordinate system (usually LMD coordinates)
+    calibration_points_target
+        3 Calibration points in target coordinate system (usually image/pixel coordinates)
+        Expects :class:`spatialdata.models.PointsModel` with calibration points in `x`/`y` column
+    calibration_points_source
+        3 Calibration points, matched to `calibration_points_target` in source coordinate system (usually LMD coordinates)
+        Expects :class:`spatialdata.models.PointsModel` with calibration points in `x`/`y` column
+    precision
+        Precision of affine transformation
+
+    Returns
+    -------
+    ShapesModel
+        Transformed shapes in target coordinate system
+
+    Raises
+    ------
+    AttributeError
+        Checks validity of shapes and calibration points data formats
+    """
+    ShapesModel.validate(shapes)
+    PointsModel.validate(calibration_points_source)
+    PointsModel.validate(calibration_points_target)
+
+    # Convert to numpy arrays
+    calibration_points_source = calibration_points_source[["x", "y"]].to_dask_array().compute()
+    calibration_points_target = calibration_points_target[["x", "y"]].to_dask_array().compute()
+
+    # Compute rotation (2x2) and translation (2x1) matrices
+    affine_transformation = compute_affine_transformation(
+        calibration_points_source, calibration_points_target, precision=precision
+    )
+    # Transform shapes
+    # Iterate through shapes and apply affine transformation
+    transformed_shapes = shapes["geometry"].apply(
+        lambda shape: shapely.transform(
+            shape, transformation=lambda geom: apply_affine_transformation(geom, affine_transformation)
+        )
+    )
+
+    # Reassign as DataFrame and parse with spatialdata
+    transformed_shapes = shapes.assign(geometry=transformed_shapes)
+
+    return ShapesModel.parse(transformed_shapes)
 
 
 def read_lmd(path: str, calibration_points_image: PointsModel, switch_orientation: bool = False) -> ShapesModel:
@@ -29,28 +88,21 @@ def read_lmd(path: str, calibration_points_image: PointsModel, switch_orientatio
     ShapesModel
         Transformed shapes in image coordinats
     """
+    PointsModel.validate(calibration_points_image)
+
     # Load LMD shapes with pyLMD
     lmd_shapes = pylmd.Collection()
     lmd_shapes.load(path)
 
-    # Transform to geopandas
+    # Transform to spatialdata models
     shapes = gpd.GeoDataFrame(geometry=[shapely.Polygon(shape.points) for shape in lmd_shapes.shapes])
-
-    calibration_points_lmd = gpd.GeoDataFrame(
-        data={"radius": np.ones(shape=len(lmd_shapes.calibration_points))},
-        geometry=[shapely.Point(point) for point in lmd_shapes.calibration_points],
-    )
-
-    calibration_points_image = gpd.GeoDataFrame(
-        data={"radius": np.ones(shape=len(calibration_points_image))},
-        geometry=[shapely.Point(row["x"], row["y"]) for _, row in calibration_points_image.iterrows()],
-    )
+    calibration_points_lmd = PointsModel.parse(lmd_shapes.calibration_points)
 
     if len(calibration_points_lmd) < 3:
         raise ValueError(f"Require at least 3 calibration points, but only received {len(calibration_points_lmd)}")
     if len(calibration_points_lmd) != len(calibration_points_image):
         raise ValueError(
-            f"Number of calibration points in image ({len(calibration_points_image)})must be equal to number of calibration points in LMD file ({len(calibration_points_lmd)})"
+            f"Number of calibration points in image ({len(calibration_points_image)}) must be equal to number of calibration points in LMD file ({len(calibration_points_lmd)})"
         )
 
     transformed_shapes = transform_shapes(
