+++ tests/read/shapes/test_geometry.py
@@ -1,14 +1,10 @@
-import geopandas as gpd
 import numpy as np
 import pytest
 from numpy.typing import NDArray
-from shapely import Polygon
-from spatialdata.models import PointsModel, ShapesModel
 
 from dvpio.read.shapes.geometry import (
     apply_affine_transformation,
     compute_affine_transformation,
-    transform_shapes,
 )
 
 test_cases = [
@@ -57,18 +53,3 @@ def test_apply_affine_transformation(
 ) -> None:
     target = apply_affine_transformation(query, affine_transformation)
     assert np.isclose(target, reference, rtol=0.001).all()
-
-
-def test_transform_shapes() -> None:
-    # Create data
-    calibration_points = PointsModel.parse(np.array([[0, 0], [1, 0], [0, 1]]))
-    shape = Polygon([[0, 0], [1, 1], [0, 1]])
-    shapes = ShapesModel.parse(gpd.GeoDataFrame(geometry=[shape] * 10))
-
-    # Transform
-    transformed_shapes = transform_shapes(
-        shapes=shapes, calibration_points_source=calibration_points, calibration_points_target=calibration_points
-    )
-
-    assert isinstance(transformed_shapes, gpd.GeoDataFrame)
-    assert len(transformed_shapes) == len(shapes)
